requirements.txt
openai
python-dotenv
tempfile
langchain-openai
langchain
sqlalchemy
psycopg2-binary
alembic
pydantic